+++ bower.json
@@ -1,0 +1,28 @@
+{
+  "name": "ml-gsd",
+  "version": "0.0.1",
+  "description": "Global Spectra Deconvolution",
+  "keywords": [
+    "optimization",
+    "curve-fitting",
+    "gsd"
+  ],
+  "main": [
+    "dist/global-spectra-deconvolution.js",
+    "dist/global-spectra-deconvolution.min.js"
+  ],
+  "authors": [
+    "Andres Castillo"
+  ],
+  "homepage": "https://github.com/mljs/global-spectra-deconvolution",
+  "license": "MIT",
+  "moduleType": [
+    "amd",
+    "globals",
+    "node"
+  ],
+  "ignore": [
+    "/*",
+    "!dist/*"
+  ]
+}
