+++ test/massPeakPicking.js
@@ -2,7 +2,7 @@
 
 var CC = require('chemcalc');
 var Stat = require('ml-stat');
-var gsdLight = require("../src/gsdLight");
+var peakPicking = require("../src/index");
 
 
 var spectrum=CC.analyseMF("Cl2.Br2", {isotopomers:'arrayXXYY', fwhm:0.01, gaussianWidth: 11});
@@ -18,10 +18,12 @@ var y=xy[1];
 var noiseLevel=Stat.array.median(y.filter(function(a) {return (a>0)}))*3;
 
 //console.log(y);
-var result=gsdLight(x, y, {noiseLevel: 0, functionType:"gaussian"});
-
+var result=peakPicking.gsd(x, y,  {noiseLevel: noiseLevel, minMaxRatio:0, broadRatio:0,smoothY:false});
+console.log("Before optmization");
 console.log(result);
-console.log(result.length);
+result = peakPicking.optimize(result,x,y,1,"gaussian");
+console.log("After optmization");
+console.log(result);
 
 describe.only('Check the peak picking of a simulated mass spectrum', function () {
